pyproject.toml
[project]
name = "searxng-langchain-demo"
version = "0.1.0"
description = "A demo project using SearxNG with LangChain"
authors = [{ name = "Rafsun Jany Arman", email = "armanrafsunjany@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.18",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
